+++ src/Admin.java
@@ -2,13 +2,13 @@ package src;
 
 import src.constants.UserRole;
 
-public class Visitor extends User {
-    public Visitor(String firstName, String lastName, String email) {
+public class Admin extends User {
+    public Admin(String firstName, String lastName, String email) {
         super(firstName, lastName, email);
     }
 
     @Override
     public UserRole getRole() {
-        return UserRole.VISITOR;
+        return UserRole.ADMIN;
     }
 }